requirements.txt
# RoboMaster TT 3D Simulator Dependencies

# Backend Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic>=2.5.0

# Configuration Management
PyYAML>=6.0.1

# HTTP Client for Mock Drones
httpx==0.25.2
aiohttp>=3.9.1
av>=15.0.0
numpy>=2.3.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
python-multipart==0.0.6